requirements.txt
markdownify
smolagents==1.13.0
requests
duckduckgo_search
pandas
PyMuPDF
sentence-transformers
transformers
gradio
pytz
pyyaml